Updates Enemy AI 2 - Enemy chase player
Enemy can now chase the player based on the distance between player and
enemy...

if player are too far from enemy, the enemy will slowly start looking
for the player by tracking the path that the player had taken.
if player are near the sight of the enemy, it will start to chase the
player.